+++ core.d.ts
@@ -1,8 +1,6 @@
 export { SmallReactive } from "./types/core/small-reactive";
 export { Component, OnConnect, OnDisconnect, OnShow, OnReload } from "./types/core/component";
-export { Switch } from "./types/core/switch";
 export { Directive } from "./types/core/directive";
 export { Module } from "./types/core/module";
 export { Injectable } from "./types/core/injectable";
 export { Service } from "./types/core/service";
-export { HttpRequest } from "./types/core/http-request"
