+++ core.js
@@ -1,8 +1,6 @@
 export { SmallReactive } from "./src/core/small-reactive.js";
 export { Component } from "./src/core/component.js";
-export { Switch } from "./src/core/switch.js";
 export { Directive } from "./src/core/directive.js";
 export { Module } from "./src/core/module.js";
 export { Injectable } from "./src/core/injectable.js";
 export { Service } from "./src/core/service.js";
-export { HttpRequest } from "./src/core/http-request.js";
