+++ tools.d.ts
@@ -1,0 +1,2 @@
+export { HttpRequest } from "./types/tools/http-request";
+export { Switch } from "./types/tools/switch";
